+++ frontend/portals/Accordion/styles.js
@@ -1,0 +1,5 @@
+import { css } from 'glamor';
+
+export const container = css({
+  paddingBottom: 16,
+});